requirements.txt
numpy>=1.22.3
tqdm~=4.63.0
matplotlib>=3.5.0
pillow>=9.0.1
opencv-python>=4.6.0
torch>=1.10.0
torchvision>=0.11.1

